+++ src/validators.js
@@ -6,7 +6,7 @@ import cardValidator from 'card-validator'
 import { base } from '@tokend/js-sdk'
 
 import { DocumentContainer } from '@/js/helpers/DocumentContainer'
-import { MathUtil } from './js/utils'
+import { MathUtil } from '@/js/utils'
 
 import * as validators from 'vuelidate/lib/validators'
 
@@ -67,6 +67,21 @@ export const noMoreThanAvailableOnBalance = balance => value => {
 
 export const noMoreThanAvailableForIssuance = available => value => {
   return MathUtil.compare(available, value) >= 0
+}
+
+export const moreThenMin = minValue => value => {
+  if (MathUtil.compare(minValue, value) === -1) {
+    return false
+  } else {
+    return true
+  }
+}
+export const lessThenMax = maxValue => value => {
+  if (MathUtil.compare(value, maxValue) === 1) {
+    return false
+  } else {
+    return true
+  }
 }
 
 export const maxDecimalDigitsCount = maxDecimalDigitsCount => value => {