+++ src/vue/modules/deposit-fiat-card/index.vue
@@ -29,23 +29,15 @@
       </div>
     </div>
     <div class="app__form-row">
-      <input-field
-        white-autofill
-        class="app__form-field"
-        v-model.trim="form.amount"
-        type="number"
-        :min="MIN_AMOUNT"
-        :max="MAX_AMOUNT"
-        :step="config.minAmount"
+      <amount-input-field
+        v-model="form.amount"
         name="deposit-fiat-card-amount"
-        @blur="touchField('form.amount')"
-        :error-message="getFieldErrorMessage('form.amount', {
-          maxDecimalDigitsCount: form.asset.trailingDigitsCount
-        })"
+        validation-type="outgoing"
         :label="'deposit-fiat-card-module.amount' | globalize({
           asset: form.asset.code
         })"
         :disabled="formMixin.isDisabled"
+        :asset="form.asset"
       />
     </div>
     <div class="app__form-row deposit-fiat-card__form-row">
@@ -203,7 +195,6 @@ import { api } from '@/api'
 import { MathUtil } from '@/js/utils'
 import {
   required,
-  maxDecimalDigitsCount,
   cardNumber,
   cardExpirationDate,
   cardCVV3,
@@ -295,12 +286,6 @@ export default {
     return {
       form: {
         asset: { required },
-        amount: {
-          required,
-          maxDecimalDigitsCount: maxDecimalDigitsCount(
-            this.config.decimalPoints
-          ),
-        },
         cardNumber: {
           required,
           cardNumber,