+++ src/vue/modules/issuance-form/components/create-issuance-form.spec.js
@@ -1,6 +1,7 @@
 import CreateIssuanceForm from './create-issuance-form'
 
 import Vuelidate from 'vuelidate'
+import Vuex from 'vuex'
 
 import { createLocalVue, mount, shallowMount } from '@vue/test-utils'
 
@@ -11,6 +12,15 @@ const localVue = createLocalVue()
 localVue.use(Vuelidate)
 
 describe('Create issuance form', () => {
+  const store = new Vuex.Store({
+    modules: {
+      account: {
+        getters: {
+          accountId: () => ('SOME_ACCOUNT_ID'),
+        },
+      },
+    },
+  })
   let sandbox
 
   beforeEach(() => {
@@ -25,7 +35,7 @@ describe('Create issuance form', () => {
     let wrapper
 
     beforeEach(() => {
-      wrapper = mount(CreateIssuanceForm, { localVue })
+      wrapper = mount(CreateIssuanceForm, { localVue, store })
     })
 
     const expectedResults = {
@@ -61,16 +71,22 @@ describe('Create issuance form', () => {
   })
 
   describe('created hook', () => {
-    it('sets first element of ownedAssets to form.asset property', () => {
-      const wrapper = shallowMount(CreateIssuanceForm, {
-        localVue,
-        propsData: {
-          ownedAssets: ['USD', 'BTC'],
-        },
-      })
+    it('sets first element of ownedAssets to form.asset property and calls loadFees method',
+      async () => {
+        sandbox.stub(CreateIssuanceForm.methods, 'loadFees').resolves()
 
-      expect(wrapper.vm.form.asset).to.equal('USD')
-    })
+        const wrapper = await shallowMount(CreateIssuanceForm, {
+          localVue,
+          store,
+          propsData: {
+            ownedAssets: ['USD', 'BTC'],
+          },
+        })
+
+        expect(CreateIssuanceForm.methods.loadFees).to.have.been.calledOnce
+        expect(wrapper.vm.form.asset).to.equal('USD')
+      }
+    )
   })
 
   describe('component', () => {
@@ -78,31 +94,61 @@ describe('Create issuance form', () => {
 
     beforeEach(() => {
       sandbox.stub(CreateIssuanceForm, 'created')
-      wrapper = shallowMount(CreateIssuanceForm, { localVue })
+      wrapper = shallowMount(CreateIssuanceForm, { localVue, store })
     })
 
-    describe('submit', () => {
-      it('calls proper methods and emits proper event', async () => {
-        sandbox.stub(wrapper.vm, 'createIssuance').resolves()
-        sandbox.stub(Bus, 'success')
-        sandbox.stub(wrapper.vm, 'hideConfirmation')
+    describe('method', () => {
+      describe('submit', () => {
+        it('calls proper methods and emits proper event', async () => {
+          sandbox.stub(wrapper.vm, 'createIssuance').resolves()
+          sandbox.stub(Bus, 'success')
+          sandbox.stub(wrapper.vm, 'hideConfirmation')
 
-        await wrapper.vm.submit()
+          await wrapper.vm.submit()
 
-        expect(wrapper.vm.createIssuance)
-          .to.have.been.calledOnce
-        expect(Bus.success).to.have.been.calledOnce
-        expect(wrapper.emitted('submit')).to.exist
-        expect(wrapper.vm.hideConfirmation).to.have.been.calledOnce
+          expect(wrapper.vm.createIssuance)
+            .to.have.been.calledOnce
+          expect(Bus.success).to.have.been.calledOnce
+          expect(wrapper.emitted('submit')).to.exist
+          expect(wrapper.vm.hideConfirmation).to.have.been.calledOnce
+        })
+
+        it('handles a thrown error properly', async () => {
+          sandbox.stub(wrapper.vm, 'createIssuance').rejects()
+          sandbox.stub(ErrorHandler, 'process')
+
+          await wrapper.vm.submit()
+
+          expect(ErrorHandler.process).to.have.been.calledOnce
+        })
       })
 
-      it('handles a thrown error properly', async () => {
-        sandbox.stub(wrapper.vm, 'createIssuance').rejects()
-        sandbox.stub(ErrorHandler, 'process')
+      describe('loadFees', () => {
+        it('calls calculateFees method and sets isFeesLoaded property', async () => {
+          sandbox.stub(wrapper.vm, 'calculateFees').resolves({
+            totalFee: {
+              fixed: '0',
+              calculatedPercent: '0',
+            },
+          })
 
-        await wrapper.vm.submit()
+          await wrapper.vm.loadFees()
 
-        expect(ErrorHandler.process).to.have.been.calledOnce
+          expect(wrapper.vm.calculateFees).to.have.been.calledOnce
+          expect(wrapper.vm.isFeesLoaded).to.be.true
+        })
+
+        it('handles a thrown error properly', async () => {
+          sandbox.stub(wrapper.vm, 'calculateFees').rejects()
+          sandbox.stub(ErrorHandler, 'processWithoutFeedback')
+
+          await wrapper.vm.loadFees()
+
+          expect(ErrorHandler.processWithoutFeedback)
+            .to.have.been.calledOnce
+          expect(wrapper.vm.isFeesLoaded).to.be.false
+          expect(wrapper.vm.isFeesLoadFailed).to.be.true
+        })
       })
     })
   })