+++ src/vuex/assets.module.js
@@ -48,6 +48,36 @@ export const getters = {
   [vuexTypes.assets]: state => state.assets.map(a => new AssetRecord(a)),
   [vuexTypes.assetByCode]: (_, getters) => assetCode =>
     getters[vuexTypes.assets].find(item => item.code === assetCode),
+
+  [vuexTypes.balancesAssets]: (a, getters, b, rootGetters) => {
+    return rootGetters[vuexTypes.accountBalances]
+      .map(item => item.asset)
+  },
+  [vuexTypes.fiatAssets]: (a, getters, b, rootGetters) =>
+    rootGetters[vuexTypes.accountBalances]
+      .map(item => item.asset)
+      .filter(item => item.isBaseAsset),
+  [vuexTypes.depositableAssets]: (a, getters, b, rootGetters) =>
+    rootGetters[vuexTypes.accountBalances]
+      .map(item => item.asset)
+      .filter(item => item.isDepositable),
+  [vuexTypes.coinpaymentsAssets]: (a, getters, b, rootGetters) =>
+    rootGetters[vuexTypes.accountBalances]
+      .map(item => item.asset)
+      .filter(item => item.isCoinpayments),
+  [vuexTypes.transferableBalancesAssets]: (a, getters, b, rootGetters) =>
+    rootGetters[vuexTypes.accountBalances]
+      .map(item => item.asset)
+      .filter(item => item.isTransferable),
+  [vuexTypes.withdrawableBalancesAssets]: (a, getters, b, rootGetters) =>
+    rootGetters[vuexTypes.accountBalances]
+      .map(item => item.asset)
+      .filter(item => item.isWithdrawable),
+  [vuexTypes.statsQuoteAsset]: (a, getters, b, rootGetters) =>
+    rootGetters[vuexTypes.assets]
+      .filter(item => {
+        return item.isStatsQuoteAsset
+      })[0] || {},
 }
 
 export default {